pyproject.toml
[project]
name = "genesis"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "bedrock-agentcore",
    "bedrock-agentcore-starter-toolkit",
    "pip",
    "pyyaml",
    "strands-agents[otel]",
    "strands-agents-tools",
    "boto3",
    "botocore",
    "duckduckgo-search",
    "aws-opentelemetry-distro==0.10.0",
    "python-dotenv",
    "zmq>=0.0.0",
    "openai-agents>=0.2.3",
    "plotly>=6.2.0",
    "tavily-python>=0.7.11",
    "google-adk>=1.12.0",
    "a2a-sdk>=0.3.2",
    "nest-asyncio>=1.6.0",
    "h2>=4.2.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["genesis_agent_runtime.py"]
